% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr.R
\name{abbr}
\alias{abbr}
\title{Abbreviate Strings}
\usage{
abbr(
  x,
  method = c("cds", "cd", "cs", "c"),
  rm = NULL,
  keep = NULL,
  sub = c(1L, 100000L),
  unique = TRUE
)
}
\arguments{
\item{x}{input vector of strings.}

\item{method}{abbreviation method. \code{'c'} stands for keeping capital letters;
\code{'d'} stands for digits; \code{'s'} stands for suffix including 'ytd', 'YoY', etc.}

\item{rm}{characters represented by regular expression to be removed.}

\item{keep}{characters represented by regular expression to be kept.}

\item{sub}{a 2-element vector containing the starting and ending position
of the sub-string to be extracted.}

\item{unique}{if \code{TRUE}, each abbreviated strings will be guaranteed unique.
Duplications will be suffixed by dots.}
}
\value{
a vector of abbreviated strings
}
\description{
Abbreviate Strings
}
