% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_list_elements.R
\name{all_list_elements_null_or_df}
\alias{all_list_elements_null_or_df}
\title{Are all elements in a list null or data.frames?}
\usage{
all_list_elements_null_or_df(x)
}
\arguments{
\item{x}{An object}
}
\description{
Are all elements in a list null or data.frames?
}
\examples{
all_list_elements_null_or_df(data.frame())
all_list_elements_null_or_df(list(data.frame()))
all_list_elements_null_or_df(list(1, NULL))
all_list_elements_null_or_df(list(data.frame(), NULL))
all_list_elements_null_or_df(list("a"=1, 2))
}
