% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_list_elements.R
\name{all_list_elements_null_or_fully_named_list}
\alias{all_list_elements_null_or_fully_named_list}
\title{Are all elements in a list null or fully named lists?}
\usage{
all_list_elements_null_or_fully_named_list(x)
}
\arguments{
\item{x}{An object}
}
\description{
Are all elements in a list null or fully named lists?
}
\examples{
all_list_elements_null_or_fully_named_list(data.frame())
all_list_elements_null_or_fully_named_list(list(data.frame()))
all_list_elements_null_or_fully_named_list(list(1, NULL))
all_list_elements_null_or_fully_named_list(list(list(), NULL))
all_list_elements_null_or_fully_named_list(list(list("a" = 1), NULL))
all_list_elements_null_or_fully_named_list(list("a"=1, 2))
}
