% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csutil-package.R
\docType{package}
\name{csutil-package}
\alias{csutil}
\alias{csutil-package}
\title{csutil: Utility Functions for Macroeconomic Time Series}
\description{
A collection of utility function to work with China's macro time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zemw/csutil}
  \item Report bugs at \url{https://github.com/zemw/csutil/issues}
}

}
\keyword{internal}
