% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{date2Num}
\alias{date2Num}
\alias{num2Date}
\title{Numeric Representation of Dates}
\usage{
date2Num(t)

num2Date(x)
}
\arguments{
\item{t}{An object of class \code{Date}}

\item{x}{A number representing a calendar date}
}
\value{
Returns the numeric representation of a calendar date or vice versa.
}
\description{
Converts between \link[base:Dates]{base::Date} objects and the numeric representations.
The numeric representation differs from the POSIX representation for
the benefit of readability.
For example, monthly date points in the year 2000 is represented as
2000, 2000.083, 2000.167, etc.
Each month is converted to a fraction of 1/12 between integer years, and
each day is a fraction of 1/365.
The representation is designed to work with \link[zoo:yearmon]{zoo::yearmon} and \link[zoo:yearqtr]{zoo::yearqtr}.
}
