% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{dualplot}
\alias{dualplot}
\title{Dual-Axis Plot}
\usage{
dualplot(y1, y2, ...)
}
\arguments{
\item{y1}{\code{ts} series (left)}

\item{y2}{\code{ts} series (right)}

\item{...}{parameters of \code{\link[plotrix:twoord.plot]{plotrix::twoord.plot()}}}
}
\value{
nil
}
\description{
A wrapper over \code{\link[plotrix:twoord.plot]{plotrix::twoord.plot()}} to plot two time series
on different axis.
}
