% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extend}
\alias{extend}
\title{Extend Time Series}
\usage{
extend(x, ...)
}
\arguments{
\item{x}{the time series to be extended.}

\item{...}{time series objects to be extended to \code{x}.}
}
\value{
The extended series. Note that leading or trailing \code{NA}s will be removed.
}
\description{
Extend one time series with another. This is a wrapper over \link[zoo:merge.zoo]{zoo::rbind.zoo}.
The difference is \code{rbind.zoo} will report error on overlapped indexes,
while this function will ignore the overlapped indexes on the right-hand-side
time series.
}
