% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceic.R
\name{grabTS}
\alias{grabTS}
\title{Compile Time Series Data}
\usage{
grabTS(df, out = c("df", "ts", "zoo"))
}
\arguments{
\item{df}{the dataframe returned from \code{\link[=parse2DF]{parse2DF()}}}

\item{out}{the output object type}
}
\value{
an object of the specified type
}
\description{
Compile the information contained in the dataframe returned from \code{\link[=parse2DF]{parse2DF()}}
into operational time series data (as \code{data.frame}, \code{ts}, or \code{zoo}).
}
\details{
It is strongly recommended that the dataframe only contains time series of
the same frequency. Though mixed frequencies are also supported.
Only monthly/quarterly/yearly series are supported.
Unless the output object is \code{ts}, it is not guaranteed that the returned
series are regular.
}
\examples{
\dontrun{
df = parse2DF("data-raw/testseries.m")
df_q = dplyr::filter(frequency == "Quarterly")
df_q.ts = grabTS(out = "ts")
}
}
