% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interpGR}
\alias{interpGR}
\title{Interpolate with Growth Rates}
\usage{
interpGR(
  x,
  r,
  lag = frequency(x),
  direction = c("both", "forward", "backward")
)
}
\arguments{
\item{x}{the level series.}

\item{r}{the growth rates (with the same length as \code{x})}

\item{lag}{number of lags when computing the growth rate. Default is the
frequency of the input series, i.e. \code{r} is the year-over-year growth rates.
\code{1} will indicate the growth rates are computed on two consecutive observations.}

\item{direction}{the direction in which to fill missing values.}
}
\value{
New series with missing values deducted from growth rates.
}
\description{
Interpolate or extrapolate missing values in a time series from its
corresponding growth rates.
}
