% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple Time Series}
\usage{
multiplot(y, ...)
}
\arguments{
\item{y}{an \code{mts} object}

\item{...}{parameters of \link[zoo:plot.zoo]{zoo::plot.zoo}}
}
\description{
A wrapper over \link[zoo:plot.zoo]{zoo::plot.zoo} to plot multiple time series.
}
\seealso{
\link[zoo:plot.zoo]{zoo::plot.zoo}
}
