% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{na.pad}
\alias{na.pad}
\title{Padding Time Series with \code{NA}s}
\usage{
na.pad(x, index = NULL, n = 0, sides = c("left", "right"))
}
\arguments{
\item{x}{a univariate time series.}

\item{index}{pad the time series to have the same time index as provided.
If this argument is provided, other arguments will be ignored.}

\item{n}{the total length of the series after padding}

\item{sides}{insert \code{NA}s on the left or right side.}
}
\value{
The padded series with the provided time index or the target length.
}
\description{
Pad time series with leading or trailing \code{NA}s
}
