% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceic.R
\name{parse2DF}
\alias{parse2DF}
\title{Parse CEIC file into data.frame}
\usage{
parse2DF(path, jsonData = TRUE)
}
\arguments{
\item{path}{path to the MATLAB file}

\item{jsonData}{stores the time series in JSON format}
}
\value{
a dataframe contains the description as well as the JSON-encoded
data for each series in the data file
}
\description{
Parse the MATLAB data downloaded from CEIC into a dataframe
}
