% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceic.R
\name{parse2TS}
\alias{parse2TS}
\title{Parse CEIC file into time series data}
\usage{
parse2TS(path, out = c("df", "ts", "zoo"))
}
\arguments{
\item{path}{path to the MATLAB file}

\item{out}{the output object type}
}
\value{
an object of the specified type
}
\description{
Parse the MATLAB data downloaded from CEIC into the specified time series
object (\code{data.frame}, \code{ts}, or \code{zoo}). See \code{\link[=grabTS]{grabTS()}} for details.
}
