% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas.R
\name{seasplot}
\alias{seasplot}
\title{Plot STL-seasonally-adjusted series}
\usage{
seasplot(x, ...)
}
\arguments{
\item{x}{original univariate time series}

\item{...}{arguments to be passed to \link[stats:stl]{stats::stl}}
}
\value{
An \code{ggplot} object.
}
\description{
The original series will be plotted together with the seasonally-adjusted
series and the trend.
}
