% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simAR1}
\alias{simAR1}
\title{Simulate an AR(1) process}
\usage{
simAR1(n, ar = 0.5, sd = 1, freq = 1, start = 2000)
}
\arguments{
\item{n}{number of observations}

\item{ar}{the AR(1) coefficient}

\item{sd}{standard deviation of the noise}

\item{freq}{frequency of the time series}

\item{start}{time of the first observation}
}
\value{
an \code{ts} object.
}
\description{
Simulate an AR(1) process
}
