% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_equal.R
\name{split_equal}
\alias{split_equal}
\title{Split a vector into a list with equal sized elements}
\usage{
split_equal(x, size = 10)
}
\arguments{
\item{x}{A vector}

\item{size}{Size of each block}
}
\description{
Split a vector into a list with equal sized elements
}
\examples{
split_equal(c(1:11), size = 3)
}
