% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas.R
\name{stlseas}
\alias{stlseas}
\title{Seasonal Adjustment with STL}
\usage{
stlseas(x, ...)
}
\arguments{
\item{x}{univariate time series to be adjusted}

\item{...}{arguments to be passed to \link[stats:stl]{stats::stl}}
}
\value{
the seasonally-adjusted series.
}
\description{
Seasonal Adjustment with STL
}
