% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{trans.ts}
\alias{trans.mts}
\title{Time Series Transformation}
\usage{
trans(x, ...)

\method{trans}{ts}(
  x,
  freq = c("asis", "y", "q", "m"),
  agg = c("avg", "sum", "last"),
  unit = 1,
  disYTD = FALSE,
  naFill = c("none", "interp", "locf", "ma", "kalman"),
  naReg = numeric(0),
  naYoY = numeric(0),
  seasAdj = c("none", "x11", "seats"),
  hpFilter = c("none", "trend", "cycle"),
  chg = c("asis", "log", "diff", "ld", "pct", "yoy", "idx", "ttm"),
  ...
)

\method{trans}{mts}(x, ..., colNames = NULL)
}
\arguments{
\item{x}{time series object}

\item{...}{arguments used to transform each single time series}

\item{freq}{frequency of the output series. If the intended frequency is
lower than the original one, aggregating method will apply. Transforming
to higher frequencies will generate NAs in the output series.}

\item{agg}{aggregating method for frequency transformation.}

\item{unit}{a single number to scale up or down the values.}

\item{disYTD}{set to \code{TRUE} for YTD series to convert it back to monthly
or quarterly level series.}

\item{naFill}{methods to impute missing values (see package \code{\link{imputeTS}})}

\item{naReg}{impute missing values by a set of regressors. Can be a vector
or a matrix. Must be the same length as the input time series.}

\item{naYoY}{impute missing values by YoY growth rates. Must be a vector
(in percentage) with the same length as the input time series.}

\item{seasAdj}{seasonal adjusting methods (see package \code{\link{seasonal}})}

\item{hpFilter}{applying (boosted) HP filter (see package \code{\link{bHP}})}

\item{chg}{transform the output series.
\code{log} for logarithmic transformation;
\code{diff} for first-order difference;
\code{ld} stands for log-difference;
\code{pct} computes the percentage change from a period ago;
\code{yoy} computes the percentage change from a year ago;
\code{idx} indexify the time series with the first non-NA observation set to 1;
\code{ttm} computes the rolling sum of past 1 year.}

\item{colNames}{rename the columns. Must be the same length as the number
of columns of the input series.}
}
\value{
the transformed series. Output series may differ in length
depending on the transformation.
}
\description{
Performs time series transformation including aggregation, missing-value
imputation, seasonal adjustment and so on. Transformations are executed
sequentially as ordered in the argument list.

For multiple time series input, the same transformation will be applied
for each series.
}
