% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tsapply}
\alias{tsapply}
\title{Apply Function to Each Time Series}
\usage{
tsapply(x, FUN, ...)
}
\arguments{
\item{x}{an \code{mts} object}

\item{FUN}{the function to be applied}

\item{...}{further arguments to be passed to \code{lapply}}
}
\description{
A wrapper over \link[base:lapply]{base::lapply}, which applies function to each column
of an \code{mts} object, and returns the updated \code{mts} object.
}
