% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ts2DF}
\alias{ts2DF}
\alias{df2TS}
\title{Coercion between \code{ts} and \code{date.frame}}
\usage{
ts2DF(x)

df2TS(df, dateIndex = 1)
}
\arguments{
\item{x}{An object of class \code{ts}}

\item{df}{A \code{data.frame} containing time series data}

\item{dateIndex}{Index of the \code{Date} column}
}
\value{
The time series data in \code{data.frame} or \code{ts}.
}
\description{
Methods to coercing between \link[stats:ts]{stats::ts} and \link[base:data.frame]{base::data.frame} representation
of time series data. The time index is converted to numbers (see \link{date2Num})
when coercing to \code{ts} objects, and \code{Date} when coercing to \code{data.frame}.
}
