% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tsdiff}
\alias{tsdiff}
\title{Lagged Differences}
\usage{
tsdiff(x, lag = 1, na.pad = FALSE)
}
\arguments{
\item{x}{input time series to be differenced.}

\item{lag}{an integer indicating which lag to use.}

\item{na.pad}{whether to pad the output series to have the same
length as the input series.}
}
\value{
the differenced series.
}
\description{
Lagged Differences for \code{ts} objects with \code{NA} padding.
}
