% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tsgrowth}
\alias{tsgrowth}
\title{Compute Growth Rate}
\usage{
tsgrowth(x, lag = 1, na.pad = FALSE, pct = TRUE)
}
\arguments{
\item{x}{input level series.}

\item{lag}{an integer indicating which lag to use.}

\item{na.pad}{whether to pad the output series to have the same
length as the input series.}

\item{pct}{output in percentage points.}
}
\value{
the growth rate series.
}
\description{
Compute Growth Rate
}
