% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_dfs_within_fully_named_list.R
\name{unnest_dfs_within_fully_named_list}
\alias{unnest_dfs_within_fully_named_list}
\title{Unnest data.frames within fully named list}
\usage{
unnest_dfs_within_fully_named_list(x)
}
\arguments{
\item{x}{A list of data.frames}
}
\description{
Unnest data.frames within fully named list
}
\examples{
x <- list(
  list(
    "a" = data.frame("v1"=1),
    "b" = data.frame("v2"=3)
  ),
  list(
    "a" = data.frame("v1"=10),
    "b" = data.frame("v2"=30),
    "d" = data.frame("v3"=50)
  )
)
print(x)
unnest_dfs_within_fully_named_list(x)
}
