% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas.R
\name{x13seas}
\alias{x13seas}
\title{Seasonal Adjustment with X13}
\usage{
x13seas(x, x11 = FALSE, cny = NULL, ..., final = TRUE)
}
\arguments{
\item{x}{input time series}

\item{x11}{whether to use X11 or SEATS (default)}

\item{cny}{parameters for the Chinese New Year holiday. If set to \code{NULL}
(default), no holiday effect will be adjusted. Otherwise, a pair of
integers indicating the start and end point of the holiday around
the New Year's Day.}

\item{...}{other arguments to be passed to \link[seasonal:seas]{seasonal::seas}}

\item{final}{whether to return the finalized (seasonally-adjusted)
series or the object of class "seas".}
}
\value{
Returns the seasonally-adjusted series if \code{final} is set to \code{TRUE};
otherwise the object of class "seas".
}
\description{
A wrapper over \link[seasonal:seas]{seasonal::seas} adjusting the Chinese New Year effect.
}
