% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{yearapply}
\alias{yearapply}
\title{Apply Function within a Year}
\usage{
yearapply(x, FUN)
}
\arguments{
\item{x}{the input series (\code{ts}-convertible object)}

\item{FUN}{the function to be applied}
}
\value{
Always return an \code{ts} object. Depending on the function provided,
the returned series may not have the same length as the input one.
}
\description{
Apply a function to quarterly or monthly series on yearly basis.
That is, the function will be applied only with observations within a year
and then move on to the next year.
Useful for computing year-to-date (YTD) accumulating sums or differencing
YTD sums.
}
