% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ytddiff}
\alias{ytddiff}
\title{Differencing YTD series}
\usage{
ytddiff(x, split = c("default", "avg"))
}
\arguments{
\item{x}{the input series (\code{ts} alike object)}

\item{split}{methods to split the values of the first two months.
Due to the Chinese New Year effect, some series are released with only 11
observations in a year in which the first two months are reported together.
Specify \code{avg} to split the sum of the first two months evenly between
the two months. \code{default} will leave the value for each month as it is.}
}
\value{
The differenced series.
}
\description{
Differencing year-to-date (YTD) series into quarterly or monthly levels.
}
