% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fn_via_hash_table.R
\name{apply_fn_via_hash_table}
\alias{apply_fn_via_hash_table}
\title{Apply a function via hash table}
\usage{
apply_fn_via_hash_table(x, fn, ...)
}
\arguments{
\item{x}{A vector data that needs a function applied to it.}

\item{fn}{A function that will be applied to x.}

\item{...}{Arguments that will be passed to `fn`.}
}
\description{
This function extracts the unique input values, applies the given function to
it to create a hash table (containing unique input/output combinations), and
then matches the original input to the hash table to obtain the desired output.
}
\details{
This can dramatically speed up computation if there is a lot of data and
a limited amount of unique values.
}
