% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csutil-package.R
\docType{package}
\name{csutil-package}
\alias{csutil}
\alias{csutil-package}
\title{csutil: Utility Functions for China Macro Time Series}
\description{
A collection of utility function to work with China's macro time series.
}
\keyword{internal}
