% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_list_elements.R
\name{is_all_list_elements_null_or_list}
\alias{is_all_list_elements_null_or_list}
\title{Are all elements in a list null or lists?}
\usage{
is_all_list_elements_null_or_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
Boolean.
}
\description{
Checks if A) `x` is a list, B) All elements in `x` are either null or list.
}
\examples{
is_all_list_elements_null_or_list(data.frame())
is_all_list_elements_null_or_list(list(data.frame()))
is_all_list_elements_null_or_list(list(1, NULL))
is_all_list_elements_null_or_list(list(list(), NULL))
is_all_list_elements_null_or_list(list("a"=1, 2))
}
\seealso{
Other list predicates:
\code{\link[=is_all_list_elements_null_or_df]{is_all_list_elements_null_or_df()}},
\code{\link[=is_all_list_elements_null_or_fully_named_list]{is_all_list_elements_null_or_fully_named_list()}},
\code{\link[=is_fully_named_list]{is_fully_named_list()}}
}
\concept{list predicates}
