% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_list_elements.R
\name{is_all_list_elements_null_or_list}
\alias{is_all_list_elements_null_or_list}
\title{Are all elements in a list null or lists?}
\usage{
is_all_list_elements_null_or_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks two things: that `x` is a list, and that all elements in `x` are
`NULL` or lists.
}
\examples{
is_all_list_elements_null_or_list(data.frame())
is_all_list_elements_null_or_list(list(data.frame()))
is_all_list_elements_null_or_list(list(1, NULL))
is_all_list_elements_null_or_list(list(list(), NULL))
is_all_list_elements_null_or_list(list("a"=1, 2))
}
\seealso{
\code{vignette("csutil", package = "csutil")}, which runs all four
  predicates on small example lists.

Other list predicates:
\code{\link[=is_all_list_elements_null_or_df]{is_all_list_elements_null_or_df()}},
\code{\link[=is_all_list_elements_null_or_fully_named_list]{is_all_list_elements_null_or_fully_named_list()}},
\code{\link[=is_fully_named_list]{is_fully_named_list()}}
}
\concept{list predicates}
