% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_list_elements.R
\name{is_fully_named_list}
\alias{is_fully_named_list}
\title{Is this a fully named list?}
\usage{
is_fully_named_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
Boolean.
}
\description{
Checks if `x` is a list with each element named.
}
\examples{
is_fully_named_list(list())
is_fully_named_list(list(1))
is_fully_named_list(list("a"=1))
is_fully_named_list(list("a"=1, 2))
}
\seealso{
Other list predicates:
\code{\link[=is_all_list_elements_null_or_df]{is_all_list_elements_null_or_df()}},
\code{\link[=is_all_list_elements_null_or_fully_named_list]{is_all_list_elements_null_or_fully_named_list()}},
\code{\link[=is_all_list_elements_null_or_list]{is_all_list_elements_null_or_list()}}
}
\concept{list predicates}
