% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stlseas}
\alias{stlseas}
\title{Seasonal Adjustment with STL}
\usage{
stlseas(
  x,
  s.window = 7,
  lny = c("trim", "keep", "discard"),
  nsigma = 3,
  simplify = T,
  na.action = NULL,
  ...
)
}
\arguments{
\item{x}{A univariate time series to be processed.}

\item{s.window}{The loess window for seasonal extraction to be passed on to \link[stats:stl]{stats::stl}.
Should be odd and at least 7. A default value is provided though a user-specified
value is strongly recommended.}

\item{lny}{Methods to deal with Lunar New Year effect. Default is \code{trim}, which
discards the blips in Jan and Feb if within a given number of standard deviations.
\code{keep} will keep all the blips. \code{discard} removes all blips and assumes Jan and Feb
observations align perfectly with the trend.}

\item{nsigma}{The number of standard deviations if \code{trim} is specified in \code{lny}.}

\item{simplify}{If \code{TRUE}, only the seasonally adjusted series is returned.
Otherwise, a matrix with the decomposed components will be returned.}

\item{na.action}{Function to deal with \code{NA}s in \code{x}. If \code{NULL}, linear
interpolation will be applied.}

\item{...}{Further arguments to be passed to \link[stats:stl]{stats::stl}.}
}
\value{
The seasonally-adjusted series with \code{NA}s interpolated if \code{simplify}
is \code{TRUE}; or a matrix of STL components if \code{simplify} is \code{FALSE}.
}
\description{
A seasonal adjustment procedure based on STL agnostic to the exact occasion
of the Lunar New Year (LNY) holiday. The LNY occurs in January or February
arbitrarily, which causes blips in Jan-Feb observations even after seasonal
smoothing. This function augments the standard procedure assuming the blips
in Jan-Feb values after subtracting the seasonal component are distortions
induced by the LNY holiday, and therefore can be discarded as long as they
are not unusually large (beyond some standard deviations).
The remedy makes it possible to perform seasonal adjustment to China's time
series data without lunar calendar regressors.
}
