% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas.R
\name{stlseas}
\alias{stlseas}
\title{Seasonal Adjustment with STL}
\usage{
stlseas(x, ..., plot = FALSE)
}
\arguments{
\item{x}{a univariate time series to be adjusted}

\item{...}{arguments to be passed to \link[stats:stl]{stats::stl}}

\item{plot}{weather a visualization of the the trend-cycle
decomposition will be returned}
}
\value{
If \code{plot} is set to \code{TRUE}, a \code{ggplot} object will be returned;
otherwise, the seasonally-adjusted series.
}
\description{
Seasonal Adjustment with STL
}
