% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{trans.ts}
\alias{trans.mts}
\alias{trans.zoo}
\title{Time Series Transformation}
\usage{
trans(x, ...)

\method{trans}{ts}(
  x,
  freq = frequency(x),
  agg = c("mean", "sum", "first", "last"),
  disagg = c("mean", "sum", "first", "last"),
  multiplier = 1,
  ytd = c("default", "diff", "diff_split"),
  na_impute = c("default", "interp", "locf", "ma", "kalman"),
  na_predict = c("default", "reg", "lnreg", "yoy"),
  xreg = NULL,
  seas = c("default", "x11", "seats", "stl"),
  seas_cny = c(-7, 7),
  chg = c("default", "log", "d", "ld", "pct", "yoy", "idx", "ttm"),
  outlier = c("default", "rm", "rp"),
  na_pad = FALSE,
  ...
)

\method{trans}{mts}(x, ..., col_names = NULL)

\method{trans}{zoo}(x, ...)
}
\arguments{
\item{x}{time series object}

\item{...}{arguments used to transform each single time series}

\item{freq}{frequency of the output series. If the intended frequency is
lower than the original one, aggregating method will apply. Transforming
to higher frequencies will generate NAs in the output series.}

\item{agg}{methods to aggregate time series to lower frequency.}

\item{disagg}{methods to disaggregate time series to higher frequency
(see \code{\link{tempdisagg}}).}

\item{multiplier}{a single number to scale up or down the values.}

\item{ytd}{methods to disaggregate YTD series into monthly or quarterly
series. \code{diff} for differencing within a year. Due to the Lunar New Year
(LNY) effect. NBS usually reports January and February observations
together (as the sum of the two month). Specify \code{diff_split} to split
the sum and accredit the values to Jan and Feb evenly.}

\item{na_impute}{methods to impute missing values (see package \code{\link{imputeTS}})}

\item{na_predict}{predicts missing values in the series.
\code{reg} predicts missing values by linear regression;
\code{lnreg} runs linear regression on log levels;
\code{yoy} deducts missing levels by year-over-year growth rates.}

\item{xreg}{regressors to predict missing values. Can be a vector
or a matrix. Must be the same length as the input time series.
If \code{na_predict} is set to \code{yoy}, must be a vector of growth rates
in percentage.}

\item{seas}{seasonal adjusting methods (see package \code{\link{seasonal}}).}

\item{seas_cny}{adjusting Chinese New Year effect in seasonal adjustment.
Must be a pair of integers indicating the start and end point of the holiday
around the New Year's Day. Set to \code{NULL} to ignore the holiday effect.}

\item{chg}{final transformation for the output series.
\code{log} for logarithmic transformation;
\code{d} for first-order difference;
\code{ld} stands for log-difference;
\code{pct} computes the percentage change from a period ago;
\code{yoy} computes the percentage change from a year ago;
\code{idx} indexify the time series with the first non-NA observation set to 1;
\code{ttm} computes the rolling sum of past 1 year.}

\item{outlier}{deals with outliers. \code{rm} for removing the outliers;
\code{rp} for replacing outliers with interpolated values.}

\item{na_pad}{whether the output series will be padded with \code{NA} to
ensure it has the same length as the input series.}

\item{col_names}{rename the columns. Must be the same length as the number
of columns of the input series.}
}
\value{
Transformed series. Output series may or may not be in the same
length as the input series.
}
\description{
Performs time series transformation including aggregation, missing-value
imputation, seasonal adjustment and so on. Transformations are executed
sequentially as ordered in the argument list.

For multiple time series input, the same transformation will be applied
for each series.
}
