% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{trans.ts}
\alias{trans.mts}
\title{Time Series Transformation}
\usage{
trans(x, ...)

\method{trans}{ts}(
  x,
  freq = c("asis", "y", "q", "m"),
  agg = c("avg", "sum", "last"),
  unit = 1,
  YTD = c("none", "d", "rm", "split", "avg"),
  na_impute = c("none", "interp", "locf", "ma", "kalman"),
  na_predict = c("none", "yoy", "reg", "lnreg"),
  xreg = NULL,
  seas = c("none", "x11", "x13"),
  hp_filter = c("none", "trend", "cycle"),
  chg = c("asis", "log", "d", "ld", "pct", "yoy", "idx", "ttm"),
  outlier = c("asis", "rm", "rpl"),
  ...
)

\method{trans}{mts}(x, ..., col_names = NULL)
}
\arguments{
\item{x}{time series object}

\item{...}{arguments used to transform each single time series}

\item{freq}{frequency of the output series. If the intended frequency is
lower than the original one, aggregating method will apply. Transforming
to higher frequencies will generate NAs in the output series.}

\item{agg}{aggregating method for frequency transformation.}

\item{unit}{a single number to scale up or down the values.}

\item{YTD}{methods to disaggregate YTD series into monthly or quarterly
series. \code{d} for consecutive differencing within a year. Due to the Lunar
New Year (LNY) effect. NBS usually reports January and February observations
together (as the sum of the two month). This results in duplicated or
missing values for the first two months and distorts the normal time series.
Specify \code{rm} to remote the values for the first two months; \code{split} to split
the sum and accredit the values to Jan and Feb according to the LNY holiday
patterns; \code{avg} to split the sum equally to each month.}

\item{na_impute}{methods to impute missing values (see package \code{\link{imputeTS}})}

\item{na_predict}{predicts missing values by regression.
\code{yoy} predicts missing levels by YoY growth rates;
\code{reg} predicts missing values by linear regression;
\code{lnreg} runs linear regression on log levels.}

\item{xreg}{regressors to predict missing values. Can be a vector
or a matrix. Must be the same length as the input time series.
If \code{na_predict} is specified as \code{yoy}, must be a vector of growth rates
in percentage.}

\item{seas}{seasonal adjusting methods (see package \code{\link{seasonal}})}

\item{hp_filter}{applying (boosted) HP filter (see package \code{\link{bHP}})}

\item{chg}{transform the output series.
\code{log} for logarithmic transformation;
\code{d} for first-order difference;
\code{ld} stands for log-difference;
\code{pct} computes the percentage change from a period ago;
\code{yoy} computes the percentage change from a year ago;
\code{idx} indexify the time series with the first non-NA observation set to 1;
\code{ttm} computes the rolling sum of past 1 year.}

\item{outlier}{deals with outliers. \code{rm} for removing the outliers;
\code{rpl} for replacing outliers with interpolated values.}

\item{col_names}{rename the columns. Must be the same length as the number
of columns of the input series.}
}
\value{
Transformed series. Output series are returned in the same
length as the input series if no frequency change applied.
}
\description{
Performs time series transformation including aggregation, missing-value
imputation, seasonal adjustment and so on. Transformations are executed
sequentially as ordered in the argument list.

For multiple time series input, the same transformation will be applied
for each series.
}
